Multiple nights with enemy waves
Added a feature that allows waves to spawn over multiple nights, up to a specified maximum number of nights. Currently, if the total waves per night is not completed in time, it will carry over into the next night. The maximum waves, spawn counts, and enemy limits can also be specified for all waves on any night.